Fix half to float giving wrong results on older x86_64 CPUs on Windows
The lzcnt instruction is not supported on some older x86_64 CPUs. On these
CPUs it will silently execute a bsr instruction instead, leading to wrong
results.

Instead use the bsr instruction and one additional subtraction, which should
have a negligible impact on performance. Additionally, this likely improves
performance on ARM.

Thanks to Ray Molenkamp for tracking down this bug.